minnor refactoring
