+++ store/sled/src/lib.rs
@@ -1,3 +1,4 @@
 mod library;
+mod util;
 
 pub use crate::library::SledLibrary;
